Split "error messages"
